🔥 refactor(context-engine): remove ActiveTopicDocumentContextInjector
With doc-anchored topics, every turn in a topic is by definition about the
linked document — the chat never crosses scopes, so the per-turn `<document>` +
"use lobe-agent-documents, not PageAgent" guidance is now redundant.

Removed
- The `ActiveTopicDocumentContextInjector` provider and its Phase 4 wiring in
  `MessagesEngine`.
- `RuntimeActiveTopicDocumentContext` plus `RuntimeInitialContext.activeTopicDocument`.
- The server-side `aiAgent.execAgent` block that resolved the agent-document
  row from `appContext.documentId` to populate `activeTopicDocument`.
- The client-side `resolveActiveTopicDocumentInitialContext` resolver plus its
  unit tests and the call sites in `conversationLifecycle` / regenerate.

Kept
- `mergeAgentRuntimeInitialContexts` — still needed for mentioned-agents and
  ad-hoc tool manifest merging. Moved to a focused
  `src/store/chat/utils/runtimeInitialContext.ts` module.
- The `documentId` / `agentDocumentId` fields on `ConversationContext` — the
  downstream `lobe-agent-documents` tool calls still read them for direct row
  lookups, just without the per-turn guidance prologue.
