+++ apps/server/src/services/aiAgent/index.ts
@@ -2492,41 +2492,6 @@ export class AiAgentService {
       },
     };
 
-    if (appContext?.scope !== 'page' && appContext?.documentId) {
-      // Server is authoritative — `(agentId, documentId)` is a unique binding
-      // so a single indexed lookup both validates any caller-supplied
-      // `agentDocumentId` hint and resolves the row id when one was not
-      // provided (covers docs opened outside the active topic, e.g. skills
-      // and web docs).
-      try {
-        const row = await this.agentDocumentsService.findRowByDocumentId(
-          resolvedAgentId,
-          appContext.documentId,
-        );
-
-        initialContext = {
-          ...initialContext,
-          initialContext: {
-            activeTopicDocument: {
-              ...(row?.id ? { agentDocumentId: row.id } : {}),
-              documentId: appContext.documentId,
-              ...(row?.title ? { title: row.title } : {}),
-            },
-          },
-        };
-      } catch (error) {
-        log('execAgent: failed to resolve active topic document context: %O', error);
-        initialContext = {
-          ...initialContext,
-          initialContext: {
-            activeTopicDocument: {
-              documentId: appContext.documentId,
-            },
-          },
-        };
-      }
-    }
-
     if (appContext?.scope === 'task' && appContext.defaultTaskAssigneeAgentId) {
       initialContext = {
         ...initialContext,
