+++ packages/types/src/stepContext.ts
@@ -75,26 +75,6 @@ export interface InitialPageEditorContext {
 }
 
 /**
- * Active topic document context
- * Captured when a user leaves the topic page editor but continues the same
- * topic conversation from the regular chat surface.
- */
-export interface RuntimeActiveTopicDocumentContext {
-  /**
-   * Agent-document row ID used by lobe-agent-documents read/patch/edit APIs.
-   */
-  agentDocumentId?: string;
-  /**
-   * Underlying documents.id used by topic page routes.
-   */
-  documentId: string;
-  /**
-   * Human-readable title for model disambiguation.
-   */
-  title?: string | null;
-}
-
-/**
  * User-selected skill context for the current request
  * Captured from slash-menu skill action tags before send
  */
@@ -214,11 +194,6 @@ export interface InjectedToolManifest {
  * Remains constant throughout the operation lifecycle.
  */
 export interface RuntimeInitialContext {
-  /**
-   * Active topic document carried from page route to regular chat route.
-   * This lets the model continue document work without page-editor tools.
-   */
-  activeTopicDocument?: RuntimeActiveTopicDocumentContext;
   /**
    * Ad-hoc tool manifests injected by callers for the current request.
    * Merged into the tool resolution output without passing through ToolsEngine.
