+++ src/features/Conversation/store/slices/generation/regenerateHeteroImages.test.ts
@@ -23,9 +23,8 @@ vi.mock('@/store/chat/slices/aiChat/actions/heteroResume', () => ({
   resolveHeteroResume: () => ({ cwdChanged: false, resumeSessionId: 'sess-1' }),
 }));
 
-vi.mock('@/store/chat/utils/activeTopicDocumentContext', () => ({
+vi.mock('@/store/chat/utils/runtimeInitialContext', () => ({
   mergeAgentRuntimeInitialContexts: () => undefined,
-  resolveActiveTopicDocumentInitialContext: async () => undefined,
 }));
 
 vi.mock('@/store/chat/slices/operation/selectors', () => ({
